favourties page working
