+++ src/components/ListingsContainter.tsx
@@ -1,5 +1,4 @@
 import { FC } from 'react'
-
 import Property from './Property'
 import { ContainerProps } from '../../types/types'
